Beastiary Fixes
Fixed a bug where Beastiary pet and summoning buttons did nothing (their ids were being cast into a byte but were too high thus became negative).
Fixed some remaining language issues in the Beastiary and for Subspecies names.
Selecting a Creature Type in the Beastiary with a Creature Selected will no deselect the creature allowing the Creature Type info to be shown again.

+++ src/main/java/com/lycanitesmobs/client/gui/BaseContainerScreen.java
@@ -97,7 +97,7 @@ public abstract class BaseContainerScreen<T extends Container> extends Container
      * Called when a Button Base is pressed providing the press button's id.
      * @param buttonId The id of the button pressed.
      */
-    public abstract void actionPerformed(byte buttonId);
+    public abstract void actionPerformed(int buttonId);
 
     /**
      * Returns the font renderer.
